Make this run on Linux too for QA.

+++ src/setup.py
@@ -75,8 +75,52 @@ else:
 
 print "Building %s" % opsiclientdDescription
 
+if RUNS_ON_WINDOWS:
+	data_files = [
+		('Microsoft.VC90.MFC', glob.glob('Microsoft.VC90.MFC\\*.*')),
+		('Microsoft.VC90.CRT', glob.glob('Microsoft.VC90.CRT\\*.*')),
+		('lib\\Microsoft.VC90.MFC', glob.glob('Microsoft.VC90.MFC\\*.*')),
+		('lib\\Microsoft.VC90.CRT', glob.glob('Microsoft.VC90.CRT\\*.*')),
+		('notifier',                      [	'windows\\helpers\\notifier\\event.ini',
+							'windows\\helpers\\notifier\\action.ini',
+							'windows\\helpers\\notifier\\userlogin.ini',
+							'windows\\helpers\\notifier\\wait_for_gui.ini',
+							'windows\\helpers\\notifier\\block_login.ini',
+							'windows\\helpers\\notifier\\popup.ini',
+							'windows\\helpers\\notifier\\shutdown.ini',
+							'windows\\helpers\\notifier\\event.bmp',
+							'windows\\helpers\\notifier\\action.bmp',
+							'windows\\helpers\\notifier\\userlogin.bmp',
+							'windows\\helpers\\notifier\\wait_for_gui.bmp',
+							'windows\\helpers\\notifier\\block_login.bmp',
+							'windows\\helpers\\notifier\\popup.bmp',
+							'windows\\opsi.ico' ]),
+		('opsiclientd',                   [	'windows\\opsiclientd.conf']),
+		('locale\\de\\LC_MESSAGES',       [     '..\\gettext\\opsiclientd_de.mo']),
+		('locale\\fr\\LC_MESSAGES',       [     '..\\gettext\\opsiclientd_fr.mo']),
+		('opsiclientd\\extend.d', glob.glob('..\\extend.d\*.*')),
+	]
+else:
+	data_files = []
+data_files += tree('opsiclientd\\static_html', '..\\static_html')
 
-manifest_template = '''
+
+setup_options = {
+	"data_files": data_files,
+	"name": "opsiclientd",
+	"description": (
+		'opsiclientd is part of the desktop management solution opsi (open pc '
+		'server integration) - http://www.opsi.org'
+	),
+	"version": opsiClientDeamonVersion,
+	"url": 'http://www.opsi.org/',
+	"author": "uib GmbH <info@uib.de>",
+	"author_email": "info@uib.de",
+	"license": "GNU Affero General Public License Version 3 (AGPLv3)",
+}
+
+if RUNS_ON_WINDOWS:
+	manifest_template = '''
 <?xml version="1.0" encoding="utf-8"?>
 <assembly xmlns="urn:schemas-microsoft-com:asm.v1" manifestVersion="1.0">
   <trustInfo xmlns="urn:schemas-microsoft-com:asm.v3">
@@ -119,124 +163,75 @@ manifest_template = '''
     </compatibility>
 </assembly>
 '''
- 
-RT_MANIFEST = 24
 
-opsiclientd = Target(
-	name = "opsiclientd",
-	description = opsiclientdDescription,
-	script = "opsiclientd.py",
-	modules = ["opsiclientd"],
-	#cmdline_style='pywin32',
-	#other_resources = [(RT_MANIFEST, 1, manifest_template % dict(prog="opsiclientd"))],
-	icon_resources = [(1, "windows\\opsi.ico")]
-)
+	RT_MANIFEST = 24
 
-# manifest_template = '''
-# <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-# <assembly xmlns="urn:schemas-microsoft-com:asm.v1" manifestVersion="1.0">
-# <assemblyIdentity
-#     version="5.0.0.0"
-#     processorArchitecture="x86"
-#     name="%(prog)s"
-#     type="win32"
-# />
-# <description>%(prog)s Program</description>
-# <dependency>
-#     <dependentAssembly>
-#         <assemblyIdentity
-#             type="win32"
-#             name="Microsoft.Windows.Common-Controls"
-#             version="6.0.0.0"
-#             processorArchitecture="X86"
-#             publicKeyToken="6595b64144ccf1df"
-#             language="*"
-#         />
-#     </dependentAssembly>
-# </dependency>
-# </assembly>
-# '''
-# 
-# RT_MANIFEST = 24
+	opsiclientd = Target(
+		name="opsiclientd",
+		description=opsiclientdDescription,
+		script="opsiclientd.py",
+		modules=["opsiclientd"],
+		#cmdline_style='pywin32',
+		#other_resources = [(RT_MANIFEST, 1, manifest_template % dict(prog="opsiclientd"))],
+		icon_resources=[(1, "windows\\opsi.ico")]
+	)
 
-notifier = Target(
-	name = "notifier",
-	description = "opsi notifier",
-	script = "windows\\helpers\\notifier\\notifier.py",
-	# other_resources = [(RT_MANIFEST, 1, manifest_template % dict(prog="notifier"))],
-	icon_resources = [(1, "windows\\opsi.ico")]
-)
+	notifier = Target(
+		name = "notifier",
+		description = "opsi notifier",
+		script = "windows\\helpers\\notifier\\notifier.py",
+		# other_resources = [(RT_MANIFEST, 1, manifest_template % dict(prog="notifier"))],
+		icon_resources = [(1, "windows\\opsi.ico")]
+	)
 
-opsiclientd_rpc = Target(
-	name = "opsiclientd_rpc",
-	description = "opsi client daemon rpc tool",
-	script = "windows\\helpers\\opsiclientd_rpc\\opsiclientd_rpc.py",
-	icon_resources = [(1, "windows\\opsi.ico")]
-)
+	opsiclientd_rpc = Target(
+		name = "opsiclientd_rpc",
+		description = "opsi client daemon rpc tool",
+		script = "windows\\helpers\\opsiclientd_rpc\\opsiclientd_rpc.py",
+		icon_resources = [(1, "windows\\opsi.ico")]
+	)
 
-action_processor_starter = Target(
-	name = "action_processor_starter",
-	description = "opsi action processor starter",
-	script = "windows\\helpers\\action_processor_starter\\action_processor_starter.py",
-	icon_resources = [(1, "windows\\opsi.ico")]
-)
+	action_processor_starter = Target(
+		name = "action_processor_starter",
+		description = "opsi action processor starter",
+		script = "windows\\helpers\\action_processor_starter\\action_processor_starter.py",
+		icon_resources = [(1, "windows\\opsi.ico")]
+	)
 
-network_performance = Target(
-	name = "network_performance",
-	description = "network performance",
-	script = "tests\\network_performance.py",
-	icon_resources = [(1, "windows\\opsi.ico")]
-)
+	network_performance = Target(
+		name = "network_performance",
+		description = "network performance",
+		script = "tests\\network_performance.py",
+		icon_resources = [(1, "windows\\opsi.ico")]
+	)
 
-opsiclientd_shutdown_starter = Target(
-	name = "opsiclientd_shutdown_starter",
-	description = "opsi client daemon shutdown-starter tool",
-	script = "windows\\helpers\\opsiclientd_shutdown_starter\\opsiclientd_shutdown_starter.py",
-	icon_resources = [(1, "windows\\opsi.ico")]
-)
+	opsiclientd_shutdown_starter = Target(
+		name = "opsiclientd_shutdown_starter",
+		description = "opsi client daemon shutdown-starter tool",
+		script = "windows\\helpers\\opsiclientd_shutdown_starter\\opsiclientd_shutdown_starter.py",
+		icon_resources = [(1, "windows\\opsi.ico")]
+	)
 
-data_files = [
-	('Microsoft.VC90.MFC', glob.glob('Microsoft.VC90.MFC\\*.*')),
-	('Microsoft.VC90.CRT', glob.glob('Microsoft.VC90.CRT\\*.*')),
-	('lib\\Microsoft.VC90.MFC', glob.glob('Microsoft.VC90.MFC\\*.*')),
-	('lib\\Microsoft.VC90.CRT', glob.glob('Microsoft.VC90.CRT\\*.*')),
-	('notifier',                      [	'windows\\helpers\\notifier\\event.ini',
-						'windows\\helpers\\notifier\\action.ini',
-						'windows\\helpers\\notifier\\userlogin.ini',
-						'windows\\helpers\\notifier\\wait_for_gui.ini',
-						'windows\\helpers\\notifier\\block_login.ini',
-						'windows\\helpers\\notifier\\popup.ini',
-						'windows\\helpers\\notifier\\shutdown.ini',
-						'windows\\helpers\\notifier\\event.bmp',
-						'windows\\helpers\\notifier\\action.bmp',
-						'windows\\helpers\\notifier\\userlogin.bmp',
-						'windows\\helpers\\notifier\\wait_for_gui.bmp',
-						'windows\\helpers\\notifier\\block_login.bmp',
-						'windows\\helpers\\notifier\\popup.bmp',
-						'windows\\opsi.ico' ]),
-	('opsiclientd',                   [	'windows\\opsiclientd.conf']),
-	('locale\\de\\LC_MESSAGES',       [     '..\\gettext\\opsiclientd_de.mo']),
-	('locale\\fr\\LC_MESSAGES',       [     '..\\gettext\\opsiclientd_fr.mo']),
-	('opsiclientd\\extend.d', glob.glob('..\\extend.d\*.*')),
-]
-data_files += tree('opsiclientd\\static_html', '..\\static_html')
-
-setup(
-	options = {
+	# These are options required by py2exe
+	setup_options['options'] = {
 		"py2exe": {
 			"compressed": 1,
-			#"bundle_files": 1,
 			"optimize": 2,
 			"excludes": excludes,
-			"packages": packages
+			"packages": packages + ["OPSI", "twisted"]
 		}
-	},
-	data_files = data_files,
-	zipfile = "lib/library.zip",
-	service = [ opsiclientd ],
-	console = [ network_performance, opsiclientd_shutdown_starter ],
-	windows = [ notifier, opsiclientd_rpc, action_processor_starter ],
-)
+	}
+
+	setup_options['zipfile'] = "lib/library.zip"
+	setup_options['service'] = [opsiclientd]
+	setup_options['console'] = [network_performance, opsiclientd_shutdown_starter]
+	setup_options['windows'] = [notifier, opsiclientd_rpc, 	action_processor_starter]
+else:
+	setup_options['scripts'] = []
+	setup_options['packages'] = packages
+
+setup(**setup_options)
+
 for lang in os.listdir(os.path.join("dist", "locale")):
 	dn = os.path.join("dist", "locale", lang, "LC_MESSAGES")
 	for mo in os.listdir(dn):
@@ -245,7 +240,8 @@ for lang in os.listdir(os.path.join("dist", "locale")):
 			dst = os.path.join(dn, mo.split('_%s.mo' % lang)[0] + '.mo')
 			os.rename(src, dst)
 
-os.unlink(os.path.join("dist", "w9xpopen.exe"))
+if RUNS_ON_WINDOWS:
+	os.unlink(os.path.join("dist", "w9xpopen.exe"))
 
 print "!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!"
 print "!!!   On the target machine always replace exe AND lib   !!!"
